backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic>=0.34.2
openai>=1.45.0
chromadb==0.5.5
pymupdf==1.24.10
pdfplumber==0.11.4
python-docx==1.1.2
pydantic==2.8.2
pydantic-settings==2.5.2
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
